requirements.txt
torch
torchvision
onnx
onnxruntime
numpy
opencv-python
tqdm
pyyaml